production/backend/requirements.txt
flask
Flask-Smorest
marshmallow
python-dotenv
gunicorn
flask-cors
torch
torchvision